requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
authlib==1.3.0
itsdangerous==2.1.2
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
